C:/Program Files/Git/ãexistsById()로 존재여부 í확인 뒤 적절한 오류 메세지 표시하기

+++ src/main/java/com/lhw/thymeleaf/jpaBoard/controller/ArticleController.java
@@ -43,6 +43,9 @@ public class ArticleController {
     @RequestMapping("/doDelete")
     @ResponseBody
     public String doDelete(long id) {
+        if (articleRepository.existsById(id) == false) {
+            return id + "번 게시물은 이미 삭제되었거나 존재하지 않는 게시물입니다.";
+        }
         articleRepository.deleteById(id);
 //        Article article = articleRepository.findById(id).get();
